define, use mermaid-make-browser-string function
In the interest of keeping things tidy, I wrapped existing and extended
functionality for formatting the base64-encoded JSON payload into a
separate function. It's unclear to me how long the live-editor required
a JSON object, but it does now. I initially believed the `mermaid` and
`updateEditor` fields to be required. Alas, they are not. However, this
may leave the door open a little wider to enabling additional
configuration passing or alternative live-editor target.